test(api): deduplicate PostgreSQL embedding fixtures
